backend/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
email-validator==2.3.0
fastapi==0.136.0
httpx==0.28.1
openai==2.32.0
psycopg[binary]==3.3.3
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.26
pytest==9.0.3
slowapi==0.1.9
SQLAlchemy==2.0.49
uvicorn==0.44.0
